% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abgFunctions_modCeres.R
\name{abgDecompQ}
\alias{abgDecompQ}
\title{alpha-, beta- and gamma-diversity multiplicative decomposition}
\usage{
abgDecompQ(spxp, Z = NULL, q = 2, mult = TRUE, check = TRUE)
}
\arguments{
\item{spxp}{sites (row) by species (cols) \code{matrix} with or without \code{rownames} and \code{colnames}}

\item{Z}{similarity \code{matrix} used into all functions.}

\item{q}{parameter influencing the weight of species abundances on
diversity calcualtions as per Leinster & Cobbold 2012.}

\item{mult}{determines whether beta-diversity  should be calculated using
multiplicative decomposition (the default) or as a percentage
(B = (G-mA)/G * 100), where mA is mean alpha-diversity}

\item{check}{should arguments be checked?}
}
\description{
\code{abgDecompQ}  performs a alpha, beta, gamma multiplicative decomposition
  using Leinster's diversity indices.
}
