% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksUtils.R
\name{addDietCat}
\alias{addDietCat}
\title{ADD MISSING DIET CATEGORIES (BASL NODES) TO A NETWORK}
\usage{
addDietCat(web, missingDietcat)
}
\arguments{
\item{web}{a square and named \code{matrix} representing an adjacency network}

\item{missingDietcat}{the diet categories (i.e. basal nodes) that will
be added to \code{web}}
}
\description{
ADD MISSING DIET CATEGORIES (BASL NODES) TO A NETWORK
}
