% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abgFunctions_modCeres.R
\name{BetaDisQ}
\alias{BetaDisQ}
\title{Pairwise beta-diversity calculation}
\usage{
BetaDisQ(spxp, Z = NULL, q = 2, check = TRUE, mult = TRUE)
}
\arguments{
\item{spxp}{sites (row) by species (cols) \code{matrix} with or without \code{rownames} and \code{colnames}}

\item{Z}{similarity \code{matrix} used into all functions.}

\item{q}{parameter influencing the weight of species abundances on
diversity calcualtions as per Leinster & Cobbold 2012.}

\item{check}{should arguments be checked?}

\item{mult}{determines whether beta-diversity  should be calculated using
multiplicative decomposition (the default) or as a percentage
(B = (G-mA)/G * 100), where mA is mean alpha-diversity}
}
\description{
\code{BetaDisQ} calculates the pairwise beta-diversity (minus 1) between sites of a site by species matrix according to the q parameter using the afformentionned functions
Allows a parametrization of the dominance effect
}
