% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkProperties.R
\name{calc.extThresh}
\alias{calc.extThresh}
\title{CALCULATE MIN PREY EXTINCTION THRESHOLDS}
\usage{
calc.extThresh(x, quants, out.dir, dietcat)
}
\arguments{
\item{x}{is a list of file paths of to the list of networks (full)}

\item{quants}{defines the quantile values to use}

\item{out.dir}{is the destination directory for outputs}

\item{dietcat}{diet categories (i.e., lowest trophic level nodes that are ubiquitous across all networks)}
}
\description{
CALCULATE MIN PREY EXTINCTION THRESHOLDS
}
