% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidation.R
\name{calcCrossValidMetrics}
\alias{calcCrossValidMetrics}
\title{CALCULATE VALIDATION METRICS AND CONFUSION MATRIX}
\usage{
calcCrossValidMetrics(
  samp,
  fullDT,
  origData,
  statsModel,
  origDataVars,
  level = NULL,
  cacheArgs = NULL
)
}
\arguments{
\item{samp}{the sample number to pick to use as the test data set}

\item{fullDT}{the full dataset (not necessarily the one used to fit
\code{statsModel}, which could have been a subset (e.g. fewer columns))}

\item{origData}{the data used to fit \code{statsModel}}

\item{statsModel}{the fitted model}

\item{origDataVars}{a character vector of the variables used in model fitting (including response variable and random effects.)}

\item{level}{passed to \code{gamlss:::predict}}

\item{cacheArgs}{a named \code{list} of arguments passed to \code{Cache}}
}
\value{
a list with 2 entries
}
\description{
to allow caching without digesting the large data table
}
