% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsTools.R
\name{calcMeanBEINF}
\alias{calcMeanBEINF}
\title{CALCULATION OF THE MEAN FROM A BETA-INFLATED DISTRIBUTION}
\usage{
calcMeanBEINF(mu, nu, tau)
}
\arguments{
\item{mu}{vector of values for the mu parameter Beta-inflated distribution}

\item{nu}{vector of values for the nu parameter Beta-inflated distribution}

\item{tau}{vector of values for the tau parameter Beta-inflated distribution}
}
\value{
expected mean Y
}
\description{
adapted from \code{gamlss::meanBEINF}
}
