% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkProperties.R
\name{calcNetworkBLMetricsPCA}
\alias{calcNetworkBLMetricsPCA}
\title{NETWORK METRICS PCA FUNCTION}
\usage{
calcNetworkBLMetricsPCA(metricsDT, PLOT = FALSE, dim)
}
\arguments{
\item{metricsDT}{is a data.table of metrics (columns) calculated for each network (rows)}

\item{PLOT}{logical. controls whether a PCA biplot is produced.}

\item{dim}{is used for faster caching and should be dim(metricsDT)}
}
\description{
Calculates a PCA on a matrix of network metrics usin\code{ade4::dudi.pca}.
  The function automatically excludes covariates with >0.9 correlation
}
