% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkProperties.R
\name{calcNetworkMetricsSummStats}
\alias{calcNetworkMetricsSummStats}
\title{AVERAGING NETWORK METRICS FUNCTION}
\usage{
calcNetworkMetricsSummStats(
  DT,
  stats = c("mean", "sd", "cv", "median", "min", "max"),
  id.vars,
  measure.vars = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{DT}{is a data.table of variables on which the summary stats will be calculated}

\item{stats}{is a list of summary statistics to calculate. Select from "mean", "sd", "cv",
"median", "min", "max", where "cv" is the coefficient of variation}

\item{id.vars}{character vector of grouping variable names}

\item{measure.vars}{character vector variables to summarise. If NULL then all non id.vars are used}

\item{na.rm}{is passed to summary statistics functions}
}
\description{
Calculates a set of summary statistical metrics on a data.table of
network metrics per pixel, per group of a grouping variable (or several)
}
