% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsTools.R
\name{calcP0BEINF}
\alias{calcP0BEINF}
\title{CALCULATION OF THE PROBABILITY OF ZERO FROM A ZERO- AND ONE-INFLATED BETA-INFLATED DISTRIBUTION}
\usage{
calcP0BEINF(nu, tau)
}
\arguments{
\item{nu}{vector of values for the nu parameter of the zero- and one-inflated Beta distribution (BEINF)}

\item{tau}{vector of values for the tau parameter of the BEINF}
}
\value{
expected probability of 0 (P(Y=0))
}
\description{
Follows the parameterisation used in \code{gamlss} package.
See Rigby et al. 2020 (Distributions for Modeling Location, Scale, and Shape: using GAMLSS in R)
}
