% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksBetadiversity.R
\name{calcSpaceBetaDiv}
\alias{calcSpaceBetaDiv}
\title{SPATIAL BETA-DIVERSITY WRAPPER FUNCTION}
\usage{
calcSpaceBetaDiv(
  ff,
  mode = "decomp",
  res.dir,
  out.dir,
  sampleNetworks = FALSE,
  noReps = NULL,
  networkGroups = NULL,
  parallel = FALSE,
  noCores = 2,
  toDo = "missing",
  cacheRepo = options("reproducible.cachePath")
)
}
\arguments{
\item{ff}{is the a file path to list of networks obtained from a scenario of change.}

\item{mode}{can be either "decomp", to calculate beta-diversity as an
alpha/gamma decomposition (using \code{econetwork::divPartition}), or "pairwise",
to calculate pairwise beta-diversity between networks (using
\code{econetwork::disPairwise}).}

\item{res.dir}{is the "root" directory where "scenario networks" were stored and out.dir is
were results will be placed (in a folder tree respecting scenario folders.)}

\item{sampleNetworks}{controls if beta-diversity is to be calculated for a sample of networks (provide integer)
or all networks (FALSE)}

\item{noReps}{determines number of reps for the sampling and calculating beta-diversity}

\item{networkGroups}{groups of networks within which to calculate beta-diversity. If provided,
it should be a named list of vectors of network IDs. Sampling will occur within each group, so beware
of sample size.}

\item{parallel}{activates parallel calculation of betadiv across samples. Uses future with plan(multiprocess)}

\item{noCores}{only used in 'parallel' is TRUE. defaults to 2.}

\item{toDo}{controls if beta-diversity has to be recalculated for 'all'
scenarios or just 'missing' ones}

\item{cacheRepo}{passed to \code{reproducible::Cache}.}
}
\description{
This function loads and prepares networks to calculate temporal
spp turnover (beta diversity) and saves the outputs.
beta diversity is calculated per pairs of pixels of the same location,
but different scenarios (a baseline/current scenario and a change scenario).
this function was designed to be parallelized.
}
