% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksBetadiversity.R
\name{calcTempBetaDiv.master}
\alias{calcTempBetaDiv.master}
\title{TEMPORAL BETA-DIVERSITY WRAPPER FUNCTION FOR PIXEL BY SPP MATRICES}
\usage{
calcTempBetaDiv.master(
  masterScen.file,
  masterBL.files,
  families = c("allTaxon", "bird", "mammal", "reptile", "amphibian"),
  out.dir,
  scen.dir,
  toDo = "all",
  method = "all"
)
}
\arguments{
\item{masterScen.file}{one scenario pixelXspp matrix file path, which
can be made by \code{makeAndSaveMasterScen}.}

\item{masterBL.files}{a vectorlis of baseline pixelXspp matrix files, which
can be made by \code{makeAndSaveMasterBL}.}

\item{families}{character. determines for which families beta div will be calculated
(this subsets the species). Can be all, one or a combination of
\code{c("allTaxon", "bird", "mammal", "reptile", "amphibian")}, "allTaxon" meaning
no subsetting is done and beta-diversity metrics are calculated across all
species. Defaults to all of these.}

\item{out.dir}{is were  beta-diversity tables will be saved (in a folder tree
respecting scenario folders.)}

\item{scen.dir}{is the "root" directory where "scenario networks" were stored}

\item{toDo}{controls if beta-diversity hhas to be recalculated for 'all'
scenarios or just 'missing' ones}

\item{method}{passed to \code{networkTempBetaDiv.master}}
}
\value{
the output beta-diversity file path
}
\description{
this function loads and prepares networks to
calculate temporal spp turnover (beta diversity) and saves
the outputs. Beta diversity is calculated per pairs of pixels
of the same location, but different scenarios (a baseline
scenario and a change scenario). This function was designed to be
parallelized. The user should provide one scenario file at
a time and a vector of baseline files (which will be matched to the
scenario - SDM stat. model match)
}
