% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksBetadiversity.R
\name{calcTempBetaDivBAMR}
\alias{calcTempBetaDivBAMR}
\title{----------------------------------------------
NETWORKS BETADIVERSITY
----------------------------------------------
TEMPORAL BETA-DIVERSITY PER TETRAPOD FAMILY WRAPPER FUNCTION   ---------------------------}
\usage{
calcTempBetaDivBAMR(
  ff,
  BLweb,
  res.dir,
  out.dir,
  dietcat = dietcat,
  families = c("allTaxon", "bird", "mammal", "reptile", "amphibian"),
  toDo = "all",
  method = "all",
  mode = "composition",
  cacheRepo = options("reproducible.cachePath")
)
}
\arguments{
\item{ff}{is the a file path to list of networks obtained from a scenario of change.}

\item{BLweb}{a baseline web (not a file path) to be compared with the scenario webs}

\item{res.dir}{is the directory where scenario simulation results were saved}

\item{out.dir}{is the directory where outputs will be saved}

\item{dietcat}{diet categories (i.e., lowest trophic level nodes that are ubiquitous
across all networks)}

\item{families}{determines for which families beta div will be calculated (this subsets the species)}

\item{toDo}{controls if beta-diversity has to be recalculated for 'all' networks or
just 'missing' ones.}

\item{method}{is passed to \code{networkTempBetaDiv}}

\item{mode}{is passed to \code{networkTempBetaDiv}}

\item{cacheRepo}{passed to \code{reproducible::Cache}}
}
\description{
This function loads and prepares networks to calculate temporal
spp turnover (beta diversity) for each tetrapod family separately and saves the outputs.
beta diversity is calculated per pairs of pixels of the same location,
but different scenarios (a baseline/current scenario and a change scenario).
this function was designed to be parallelelized.
}
