% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nu17May_Ceres.R
\name{CalculatePredatorOverlap}
\alias{CalculatePredatorOverlap}
\title{Network average predator overlap}
\usage{
CalculatePredatorOverlap(M)
}
\arguments{
\item{M}{a square \code{matrix} representing an adjacency newtork}
}
\description{
Network average predator overlap
}
