% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsTools.R
\name{calcVarBEINF}
\alias{calcVarBEINF}
\title{CALCULATION OF THE VARIANCE OF A ZERO- AND ONE-INFLATED BETA-INFLATED DISTRIBUTION}
\usage{
calcVarBEINF(mu, sigma, nu, tau)
}
\arguments{
\item{mu}{vector of values for the mu parameter of the zero- and one-inflated Beta distribution (BEINF)}

\item{sigma}{vector of values for the sigma parameter of the BEINF}

\item{nu}{vector of values for the nu parameter of the BEINF}

\item{tau}{vector of values for the tau parameter of the BEINF}
}
\value{
expected variance of Y
}
\description{
Follows the parameterisation used in \code{gamlss} package, using variance estimation described in
and Ferrari (2010). See also Rigby et al. 2020 (Distributions for Modeling Location,
Scale, and Shape: using GAMLSS in R)
}
