% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abgFunctions_modCeres.R
\name{chaoObjects}
\alias{chaoObjects}
\title{Data prep for diversity metrics}
\usage{
chaoObjects(spxp, phy)
}
\arguments{
\item{spxp}{sites (row) by species (cols) \code{matrix} with or without \code{rownames} and \code{colnames}}

\item{phy}{ultrametric phylogenetic tree.}
}
\description{
\code{chaoObjects} is a data preparation function.
It returns adequate arguments for \code{abgDecompQ}, \code{BetaDisQ} and
\code{divLeinster} to perform a diversity analysis using Chao's diversity index.
Warning: this formula only works with an ultrametric tree!
}
