% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialTools.R
\name{checkProjections}
\alias{checkProjections}
\title{CHECK PROJECTIONS}
\usage{
checkProjections(sfObj.list)
}
\arguments{
\item{sfObj.list}{is a list of spatial objects}
}
\description{
CHECK PROJECTIONS
}
