% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksDataPrep.R
\name{compileSDMsppRichness}
\alias{compileSDMsppRichness}
\title{COMPILE SDM PROJECTIONS PRESENCE/ABSENCES}
\usage{
compileSDMsppRichness(fileLs, fileFolder)
}
\arguments{
\item{fileLs}{is the list of file names (not paths!) to the different SDM projections}

\item{fileFolderis}{the path to the parent folder containing all files (and directories to the SDM projections)}
}
\description{
COMPILE SDM PROJECTIONS PRESENCE/ABSENCES
}
