% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialTools.R
\name{cropToStudyArea}
\alias{cropToStudyArea}
\title{CROP & MASK TO STUDY AREA}
\usage{
cropToStudyArea(study.area, tocrop, method = "bilinear")
}
\arguments{
\item{study.area}{is a \code{Raster} or \code{Spatial} object}

\item{tocrop}{a \code{Raster}}

\item{method}{passed to \code{raster::projectRaster} - might need to be changed for factors}
}
\value{

}
\description{
CROP & MASK TO STUDY AREA
}
