% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nu17May_Ceres.R
\name{density.degree.distribution}
\alias{density.degree.distribution}
\title{Density of degree distribution calculation}
\usage{
\method{density}{degree.distribution}(M, cumulative = TRUE, ...)
}
\arguments{
\item{M}{a square \code{matrix} representing an adjacency newtork}

\item{cumulative}{cumulative degree distribution?}
}
\description{
Density of degree distribution calculation
}
