% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nu17May_Ceres.R
\name{densityDegreeDistribution}
\alias{densityDegreeDistribution}
\title{Density of degree distribution calculation}
\usage{
densityDegreeDistribution(M, cumulative = TRUE, ...)
}
\arguments{
\item{M}{a square \code{matrix} representing an adjacency newtork}

\item{cumulative}{cumulative degree distribution?}

\item{...}{passed to \code{igraph::degree}}
}
\description{
Density of degree distribution calculation
}
