% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abgFunctions_modCeres.R
\name{divLeinster}
\alias{divLeinster}
\title{Site diversity according to Leinster & Cobbold 2012.}
\usage{
divLeinster(spxp, Z = NULL, q = 2, check = TRUE)
}
\arguments{
\item{spxp}{sites (row) by species (cols) \code{matrix} with or without \code{rownames} and \code{colnames}}

\item{Z}{similarity \code{matrix} used into all functions.}

\item{q}{parameter influencing the weight of species abundances on
diversity calcualtions as per Leinster & Cobbold 2012.}

\item{check}{should arguments be checked?}
}
\description{
\code{divLeinster} calculates the diversity of each site of a site by species
  matrix according to the q parameter (which provide more or less weight to
  species abundance), following Leinster & Cobbold 2012.
}
