% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkProperties.R
\name{.calcCentrality}
\alias{.calcCentrality}
\title{CALCULATE SPECIES CENTRALITY}
\usage{
.calcCentrality(
  network,
  sppExclude,
  metric = c("degree", "betweenness", "eigenvector")
)
}
\arguments{
\item{network}{a network coercible to \code{matrix} and \code{igraph}}

\item{sppExclude}{list of species to exclude the network before calculating degree
centrality (all other metrics are calculated on the full network)}

\item{metric}{centrality metric to calculate. Either "degree"
(using \code{igraph::degree(network, mode = "all", v = spp)}),
"betweenness" (using \code{igraph::betweenness(network, directed = TRUE)})
"eigenvector" centrality (using \code{eigen_centrality(graph = network, directed = FALSE, scale = FALSE)}).
(see Bauer et al 2010, Ecological Complexity).}
}
\description{
Internal function to calculate centrality of a list of species in a network
}
