% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidation.R
\name{.calcMeanBEINF}
\alias{.calcMeanBEINF}
\title{CALCULATION OF THE MEAN FROM A BETA-INFLATED DISTRIBUTION
adapted from `gamlss::meanBEINF`}
\usage{
.calcMeanBEINF(mu, nu, tau)
}
\arguments{
\item{mu, }{@param nu, @param tau are vectors of values for the mu, nu and tau
of the Beta-inflated distribution}
}
\description{
CALCULATION OF THE MEAN FROM A BETA-INFLATED DISTRIBUTION
adapted from `gamlss::meanBEINF`
}
