% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksDataPrep.R
\name{.compileMasterBL}
\alias{.compileMasterBL}
\title{internal function}
\usage{
.compileMasterBL(file.ls, dietcat)
}
\arguments{
\item{file.ls}{file list}

\item{dietcat}{diet categories (i.e., lowest trophic level nodes that are ubiquitous
across all networks)}
}
\description{
internal function
}
