% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksUtils.R
\name{.convert2igraph}
\alias{.convert2igraph}
\title{CONVERT LIST OF ADJENCY MATRICES TO IGRAPH}
\usage{
.convert2igraph(networkList)
}
\arguments{
\item{networkList}{is a list of adjacency matrices that can be converted to
\code{igraph} using \code{igraph::graph_from_adjacency_matrix}}
}
\description{
Wrapper around \code{igraph::graph_from_adjacency_matrix} that
 outputs \code{NA} for empty networks.
}
