% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hypervolumes_functions.R
\name{.HVcalc}
\alias{.HVcalc}
\title{HYPERVOLUME COMPUTATION WRAPPER}
\usage{
.HVcalc(
  big.table,
  HVpoints,
  noAxes,
  HVmethod,
  ordination,
  HVnames,
  bwHV1,
  bwHV2,
  verbose,
  ...
)
}
\arguments{
\item{big.table}{is a \code{data.table} contained the raw data for the two hypervolumes.
Used to subset \code{HVpoints} according to HVnames, so rows have to correspond}

\item{HVpoints}{\code{data.table} of points used to build hypervolumes and final number of axes.}

\item{noAxes}{determines the number of axes/columns to use for hypervolume calculation}

\item{HVmethod}{determines the method used to calculate hypervolumes - passed to \code{hypervolume::hypervolume} method argument}

\item{ordination}{determines whether hypervolumes will be calculated on the raw data ("none") of on
ordination factor scores (one of "PCA", "HillSmith", "dudi.mix"). Defaults to using a PCA.}

\item{HVnames}{\code{character} vector of length 2 of hypervolume names.}

\item{bwHV1}{determines the bandwidth value for first hypervolume. Only used if \code{freeBW} is FALSE. see \code{hypervolume::hypervolume}
Only used if \code{HVmethod} is "box" or "gaussian".}

\item{bwHV2}{determines the bandwidth value for second hypervolume. Only used if \code{freeBW} is FALSE. see \code{hypervolume::hypervolume}
Only used if \code{HVmethod} is "box" or "gaussian".}

\item{verbose}{is passed to \code{hypervolume::*} functions to control printing of diagnostic outputs}

\item{...}{further arguments passed to \code{hypervolume::hypervolume}.}
}
\value{
a \code{list} with entries:
 \itemize{
   \item 'HV1'
   \item 'HV2'
   \item 'volume.set' - the result of \code{hypervolume::hypervolume_set}
   \item 'hv.centroid.dist' and 'hv.min.dist' - the result of \code{hypervolume::hypervolume_distance(..., type = "centroid")} and \code{hypervolume::hypervolume_distance(..., type = "minimum")}
   \item 'Bandwidth' - \code{NA} if \code{HVmethod} is not "box" or "gaussian")
   \item 'HV1.disjfact' and 'HV2.disjfact' (\code{NA} if \code{HVmethod} is not "box")
   \item 'SVM_nu' and 'SVM_gama' (\code{NA} if \code{HVmethod} is "box" or "gaussian")
 }
}
\description{
Calculates two hypervolumes and compares them in terms of
  overlap and distance. When appropriate (i.e. \code{HVmethod == "box"}),
  it also calculates the disjunct factor and repeats the calculations
  with larger bandwidths (if \code{freeBW == TRUE})
}
