% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hypervolumes_functions.R
\name{.ordinationPlots}
\alias{.ordinationPlots}
\title{ORDINATION PLOT SAVING FUNCTION}
\usage{
.ordinationPlots(
  ordi,
  fscores,
  PEV,
  datatable,
  HVidvar,
  HVnames,
  ordination,
  outputs.dir,
  file.suffix
)
}
\arguments{
\item{ordi}{the ordination object}

\item{fscores}{factor scores}

\item{PEV}{percent explained variance per PC}

\item{datatable}{is a \code{data.table} contained the raw data for the two hypervolumes}

\item{HVidvar}{the name or number of the column (in \code{HVdata1/2}) containing the ID of each hypervolume.}

\item{HVnames}{character vector of hypervolume names}

\item{ordination}{determines whether hypervolumes will be calculated on the raw data ("none") of on
ordination factor scores (one of "PCA", "HillSmith", "dudi.mix"). Defaults to using a PCA.}

\item{outputs.dir}{is the directory to store results}

\item{file.suffix}{is a character string used as a suffix in file names}
}
\value{
NULL, just saves plots to a PDF
}
\description{
Calculates an ordination across two datasets prior to calculating two hypervolumes (one from each dataset)
}
