% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hypervolumes_functions.R
\name{.scaleVars}
\alias{.scaleVars}
\title{SCALE VARIABLES}
\usage{
.scaleVars(datatable, init.vars)
}
\arguments{
\item{datatable}{a table object coercible to \code{data.frame}}

\item{init.vars}{is a vector with the variable indices (column numbers) to use. Defaults to NULL
and all variables are used - ATTENTION the user should not have more than 8 variables unless
using an ordination to reduce dimensionality. See Blonder et al 2014.}
}
\value{
scaled data.frame
}
\description{
Scales variables/columns in \code{init.vars} using \code{scale(..., center = FALSE, scale = TRUE)}
Variables that only have 0s, will be assigned 1e-6
}
