% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksUtils.R
\name{findIUCNAcceptedName}
\alias{findIUCNAcceptedName}
\title{FIND SPECIES ACCEPTED NAMES (IUCN)}
\usage{
findIUCNAcceptedName(species)
}
\arguments{
\item{species}{character vector of species scientific names. see \code{rredlist::rl_search}}
}
\description{
FIND SPECIES ACCEPTED NAMES (IUCN)
}
