% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksUtils.R
\name{findIUCNStatus}
\alias{findIUCNStatus}
\title{FIND SPECIES IUCN STATUS}
\usage{
findIUCNStatus(species, region = NULL)
}
\arguments{
\item{species}{character vector of species scientific names. see \code{rredlist::rl_search}}

\item{region}{region to search in. see \code{rredlist::rl_regions}}
}
\description{
FIND SPECIES IUCN STATUS
}
