% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting2.R
\name{fitting2}
\alias{fitting2}
\title{Modified version of \code{fitting} from the \code{brainwaver} R package,
that removes unnecessary code beyond the fitting of a truncated power law
and suppresses histogram and writing outputs}
\usage{
fitting2(degree.dist)
}
\arguments{
\item{degree.dist}{see \code{brainwaver::fitting}}
}
\description{
Modified version of \code{fitting} from the \code{brainwaver} R package,
that removes unnecessary code beyond the fitting of a truncated power law
and suppresses histogram and writing outputs
}
