% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nu17May_Ceres.R
\name{FractionOfBasal}
\alias{FractionOfBasal}
\title{Proportion of basal nodes}
\usage{
FractionOfBasal(M)
}
\arguments{
\item{M}{a square \code{matrix} representing an adjacency newtork}
}
\description{
Proportion of basal nodes
}
