% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nu17May_Ceres.R
\name{FractionOfCannibalism}
\alias{FractionOfCannibalism}
\title{Proportion of cannibalistic links}
\usage{
FractionOfCannibalism(M)
}
\arguments{
\item{M}{a square \code{matrix} representing an adjacency newtork}
}
\description{
Proportion of cannibalistic links
}
