% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nu17May_Ceres.R
\name{FractionOfIntermediate}
\alias{FractionOfIntermediate}
\title{Proportion of intermediate consumers}
\usage{
FractionOfIntermediate(M)
}
\arguments{
\item{M}{a square \code{matrix} representing an adjacency newtork}
}
\description{
Proportion of intermediate consumers
}
