% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nu17May_Ceres.R
\name{FractionOfTop}
\alias{FractionOfTop}
\title{Proportion of top consumers}
\usage{
FractionOfTop(M)
}
\arguments{
\item{M}{a square \code{matrix} representing an adjacency newtork}
}
\description{
Proportion of top consumers
}
