% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksDataPrep.R
\name{fun.dbf2raster}
\alias{fun.dbf2raster}
\title{CONVERT SPECIES DATABASE FILES TO RASTERS}
\usage{
fun.dbf2raster(SPPPA, mask.dir = NULL)
}
\arguments{
\item{SPPPA}{\code{data.frame}  of species presense and absences
with firt column being the cell ID ('PAGENAME') and the second the species
presence/absence}

\item{mask.dir}{folder where study area mask rasters and dbf file with pixel IDs can be found}
}
\description{
CONVERT SPECIES DATABASE FILES TO RASTERS
}
