% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksDataPrep.R
\name{fun.load.many.rdata}
\alias{fun.load.many.rdata}
\title{READ SEVERAL RDATA FILES TO A LIST}
\usage{
fun.load.many.rdata(file.list = NULL)
}
\arguments{
\item{file.list}{list of RDATA files}
}
\description{
Author: João Braga 2016
}
