% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksDataPrep.R
\name{fun.PRESENCE.ABSENCE}
\alias{fun.PRESENCE.ABSENCE}
\title{DETECT SPECIES PRESENCES IN PIXEL BY A THRESHOLD NUMBER OF PRESENCES}
\usage{
fun.PRESENCE.ABSENCE(species.dist = NULL, threshold = NULL, opt.only = NULL)
}
\arguments{
\item{species.dist}{a species distribution dbf file}

\item{threshold}{cell habitat percentage threshold to consider a species present}

\item{opt.only}{if \code{TRUE} only presences in optimal habitats are
considered, if \code{FALSE} both secondary and optimal habitat presences are used}
}
\description{
DETECT SPECIES PRESENCES IN PIXEL BY A THRESHOLD NUMBER OF PRESENCES
}
