% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPEF_own.R
\name{getPEF.own}
\alias{getPEF.own}
\title{Modified \code{gamlss::getPEF}}
\usage{
getPEF.own(
  obj = NULL,
  term = NULL,
  data = NULL,
  n.points = 100,
  parameter = c("mu", "sigma", "nu", "tau", "all"),
  level = NULL,
  type = c("response", "link"),
  how = c("mean", "median", "last"),
  fixed.at = list(),
  plot = FALSE,
  output = c("function", "vals")
)
}
\arguments{
\item{obj}{see \code{gamlss::getPEF}}

\item{term}{see \code{gamlss::getPEF}}

\item{data}{see \code{gamlss::getPEF}}

\item{n.points}{see \code{gamlss::getPEF}}

\item{parameter}{see \code{gamlss::getPEF}. Also allows "all" to
generate functions for all parameters at once}

\item{level}{see \code{gamlss::getPEF}}

\item{type}{see \code{gamlss::getPEF}}

\item{how}{see \code{gamlss::getPEF}}

\item{fixed.at}{see \code{gamlss::getPEF}}

\item{plot}{see \code{gamlss::getPEF}}

\item{output}{chose between outputting the function or predicted values}
}
\description{
added \code{how = "mean"} and \code{level} to gamlss::getPEF
added \code{output} to chose between function or predicted values
}
