% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksBuildLocalWebs.R
\name{HabLoss_SppDist_localweb}
\alias{HabLoss_SppDist_localweb}
\title{Function to build local networks according to LC changes and changes in spp dists}
\usage{
HabLoss_SppDist_localweb(
  metaweb = NULL,
  ORIGSPP = NULL,
  SPP.HAB = NULL,
  PIX.HAB.ORIG = NULL,
  PIX.HAB.FUT = NULL,
  FUTSPP = NULL,
  EXT.TRSH = NULL,
  dietcat = NULL,
  HELP = TRUE
)
}
\arguments{
\item{metaweb}{a square \code{matrix} representing an adjacency metaweb}

\item{ORIGSPP}{species IDs of current species presences}

\item{SPP.HAB}{a spp x habitat \code{matrix}}

\item{PIX.HAB.ORIG}{baseline pixel x habitat \code{matrix}}

\item{PIX.HAB.FUT}{future pixel x habitat \code{matrix}}

\item{FUTSPP}{species IDs of future species presences}

\item{EXT.TRSH}{threshold of lost prey that determines spp extinction. Varies from 0 to 1,
increasing with predator sensitivity (e.g. 0.9 = spp dies if it is left with 90\% of orignal prey)}

\item{dietcat}{diet categories (i.e., lowest trophic level nodes that are ubiquitous
across all networks)}

\item{HELP}{print help}

\item{SPPCODE}{one or more species IDs}

\item{PIX.HAB}{a pix x habitat \code{matrix}}
}
\description{
Function to build local networks according to LC changes and changes in spp dists
}
