% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hypervolumes_functions.R
\name{hypervolumes}
\alias{hypervolumes}
\title{CALCULATE AND COMPARE TWO HYPERVOLUMES}
\usage{
hypervolumes(
  HVdata1,
  HVdata2,
  HVidvar,
  ordination = "PCA",
  init.vars = NULL,
  noAxes = NULL,
  do.scale = FALSE,
  HVmethod = "box",
  freeBW = FALSE,
  bwHV1 = NULL,
  bwHV2 = NULL,
  no.runs = 1,
  plotOrdi = TRUE,
  plotHV = TRUE,
  saveOrdi = TRUE,
  outputs.dir,
  file.suffix,
  verbose = TRUE,
  plotHVDots = NULL,
  ...
)
}
\arguments{
\item{HVdata1}{the \code{data.frame} from which the first hypervolume will be calculated}

\item{HVdata2}{the \code{data.frame} from  which the second hypervolume will be calculated.
\code{HVdata1} and \code{HVdata2} must have the same column names and ideally the same number of rows.}

\item{HVidvar}{the name or number of the column (in \code{HVdata1/2}) containing the ID of each hypervolume.}

\item{ordination}{determines whether hypervolumes will be calculated on the raw data ("none") of on
ordination factor scores (one of "PCA", "HillSmith", "dudi.mix"). Defaults to using a PCA.}

\item{init.vars}{is a vector with the variable indices (column numbers) to use. Defaults to NULL
and all variables are used - ATTENTION the user should not have more than 8 variables unless
using an ordination to reduce dimensionality. See Blonder et al 2014.}

\item{noAxes}{determines the number of axes/columns to use for hypervolume calculation}

\item{do.scale}{activates scaling prior to the ordination/calculation of HVs}

\item{HVmethod}{determines the method used to calculate hypervolumes - passed to \code{hypervolume::hypervolume} method argument}

\item{freeBW}{determines whether a bandwidth estimator will be used to calculate bandwith per variable.
Only used if \code{HVmethod} is "box" or "gaussian".}

\item{bwHV1}{determines the bandwidth value for first hypervolume. Only used if \code{freeBW} is FALSE. see \code{hypervolume::hypervolume}
Only used if \code{HVmethod} is "box" or "gaussian".}

\item{bwHV2}{determines the bandwidth value for second hypervolume. Only used if \code{freeBW} is FALSE. see \code{hypervolume::hypervolume}
Only used if \code{HVmethod} is "box" or "gaussian".}

\item{no.runs}{determines how many times the HV calculations and comparisons are repeated}

\item{plotOrdi}{activates/desactivates plotting for ordinations - plots are saved to PDFs not plotted interactively}

\item{plotHV}{activates/desactivates plotting for hypervolumes - plots are saved to PDFs not plotted interactively}

\item{saveOrdi}{activates/desactivates saving ordinations outputs}

\item{outputs.dir}{is the directory to store results}

\item{file.suffix}{is a character string used as a suffix in file names}

\item{verbose}{is passed to \code{hypervolume::*} functions to control printing of diagnostic outputs}

\item{plotHVDots}{a \code{list} of further arguments passed to \code{hypervolume:::plot.HypervolumeList}.}

\item{...}{further arguments passed to \code{hypervolume::hypervolume}.}
}
\value{
Nothing is returned. Hypervolumes and comparisons are saved \code{outputs.dir}
}
\description{
Calculates two \code{hypervolumes} on two sets of "raw" data or on the factor scores of
an ordination done on the entire dataset (across the two sets of data).
}
