% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nu17May_Ceres.R
\name{IsOmnivore2}
\alias{IsOmnivore2}
\title{Calculate if a spp is an omnivore based on the number of different trophic levels predated by a spp}
\usage{
IsOmnivore2(M, level = PreyAveragedTrophicLevel)
}
\arguments{
\item{M}{a square \code{matrix} representing an adjacency newtork}

\item{level}{a function to calculate the trophic level. See \code{cheddar}
for options}
}
\description{
Calculate if a spp is an omnivore based on the number of different trophic levels predated by a spp
}
