% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySE.R
\name{length2}
\alias{length2}
\title{New version of length which can handle NA's}
\usage{
length2(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector}

\item{na.rm}{if TRUE does not acount for NA's when taking the length of \code{x}.
Defaults to FALSE which has the same behaviour as \code{lenght}.}
}
\value{
an integer of the length of \code{x} accounting for NAs or not.
}
\description{
New version of length which can handle NA's
}
