% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksDataPrep.R
\name{loadAndSummarizeResults}
\alias{loadAndSummarizeResults}
\title{LOAD AND SUMMARISE METRICS RESULTS FUNCTION}
\usage{
loadAndSummarizeResults(...)
}
\arguments{
\item{...}{passed to \code{loadResultsMetrics}, \code{calcNetworkMetricsSummStats}
and \code{reproducible::Cache}}
}
\description{
compiles and summarises simulation network metrics results.
this function is a wrapper to \code{loadResultsMetrics} and \code{calcNetworkMetricsSummStats}
functions and passes all arguments to these functions, as well as \code{reproducible::Cache}
}
