% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialTools.R
\name{loadBindSpatialObjs}
\alias{loadBindSpatialObjs}
\title{JOIN SPATIAL OBJECTS FUNCTION}
\usage{
loadBindSpatialObjs(files, destinationPath, urls = NULL)
}
\arguments{
\item{files}{is a character string of file names}

\item{destinationPath}{passed to \code{reproducible::prepInputs}}

\item{urls}{a character vector or list of URLs for each file}
}
\value{
joined \code{shapefile} or \code{raster}
}
\description{
joins \code{shapefiles} or \code{rasters}
}
