% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksDataPrep.R
\name{loadResultsSpaceBetaDiv}
\alias{loadResultsSpaceBetaDiv}
\title{LOAD SPATIAL BETA DIVERSITY RESULTS FUNCTION}
\usage{
loadResultsSpaceBetaDiv(file.ls)
}
\arguments{
\item{file.ls}{list of files to compile}
}
\description{
compiles calculations of spatial beta diversity results and outputs the data.table
}
