% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksDataPrep.R
\name{loadResultsTempBetaDiv}
\alias{loadResultsTempBetaDiv}
\title{LOAD TEMPORAL BETA DIVERSITY RESULTS FUNCTION}
\usage{
loadResultsTempBetaDiv(file.ls)
}
\arguments{
\item{file.ls}{list of files to compile}
}
\description{
compiles calculations of temporal beta diversity results and outputs the data.table
}
