% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksDataPrep.R
\name{makeAndSaveMasterBL}
\alias{makeAndSaveMasterBL}
\title{MAKE AND SAVE BASELINE MASTER MATRICES (SPP X PIXEL MATRICES) FUNCTION
compiles all scenario networks into a data.table and saves it to an
 .rds file}
\usage{
makeAndSaveMasterBL(
  x,
  returnMaster = FALSE,
  dietcat,
  redo = FALSE,
  save = TRUE
)
}
\arguments{
\item{x}{is the .RData file containing the list of networks.}

\item{returnMaster}{controls whether the function returns the master matrix or not. Defaults
to FALSE and returns the saved file path}

\item{dietcat}{diet categories (i.e., lowest trophic level nodes that are ubiquitous across all networks)}

\item{redo}{controls whether masters will be recalculated and re-saved
in case their saved files already exist}

\item{save}{controls saving}
}
\description{
MAKE AND SAVE BASELINE MASTER MATRICES (SPP X PIXEL MATRICES) FUNCTION
compiles all scenario networks into a data.table and saves it to an
 .rds file
}
