% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksUtils.R
\name{makeBioregDT}
\alias{makeBioregDT}
\title{MAKE TABLE OF BIOGEOGRAPHIC REGION BY PIXEL ID}
\usage{
makeBioregDT(rasterToMatch = NULL, maskID = NULL, args = NULL)
}
\arguments{
\item{rasterToMatch}{must be provided and match pixel IDs and names
in \code{maskID}}

\item{maskID}{table with at least the column 'PageName' (the "names"
of non-NA pixels). NA-ed pixels in rasterToMatch should not appear
in this table.}

\item{args}{named list of other arguments passed to \code{reproducible::prepInputs}
to download and process a layer of biogeographic regions. The
following defaults are used, if not provided:
\code{url = "https://www.eea.europa.eu/data-and-maps/data/biogeographical-regions-europe-3/zipped-shapefile-format-vector-polygon/zipped-shapefile-format-vector-polygon/at_download/file"}
\code{archive = "file.zip"}
\code{targetFile = "BiogeoRegions2016.shp"}
\code{rasterToMatch = mask10k}
\code{fun = "raster::shapefile"}
\code{overwrite = TRUE}
\code{destinationPath = "Habitats/Bioregions"}
\code{useCache = FALSE}
\code{cacheRepo = options("reproducible.cachePath")}}
}
\description{
Downloads and processes biogeographic regions layer
to identify biogeographi region names per network pixel ID
}
