% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTransparent.R
\name{makeTransparent}
\alias{makeTransparent}
\title{Function  to make transparent versions of R colors.}
\usage{
makeTransparent(someColor, alpha = 100)
}
\arguments{
\item{someColor}{takes a vector of color as \code{col} in \code{col2rgb}}

\item{alpha}{determines the transparency on a 0-255 scale}
}
\description{
Function  to make transparent versions of R colors.
}
