% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nu17May_Ceres.R
\name{MaximumSimilarity}
\alias{MaximumSimilarity}
\title{Network similarity
this is more like mean similarity than "maximum"}
\usage{
MaximumSimilarity(M)
}
\arguments{
\item{M}{a square \code{matrix} representing an adjacency newtork}
}
\description{
Network similarity
this is more like mean similarity than "maximum"
}
