% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nu17May_Ceres.R
\name{MeanFoodChainLength}
\alias{MeanFoodChainLength}
\title{Network mean food chain length}
\usage{
MeanFoodChainLength(M)
}
\arguments{
\item{M}{a square \code{matrix} representing an adjacency newtork}
}
\description{
Network mean food chain length
}
