% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nu17May_Ceres.R
\name{MeanGenerality}
\alias{MeanGenerality}
\title{Network mean Generality}
\usage{
MeanGenerality(M)
}
\arguments{
\item{M}{a square \code{matrix} representing an adjacency newtork}
}
\description{
Network mean Generality
}
