% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nu17May_Ceres.R
\name{MeanGenerality_norm}
\alias{MeanGenerality_norm}
\title{Network normalised mean Generality}
\usage{
MeanGenerality_norm(M)
}
\arguments{
\item{M}{a square \code{matrix} representing an adjacency newtork}
}
\description{
Network normalised mean Generality
}
