% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nu17May_Ceres.R
\name{MeanGenerality_norm2}
\alias{MeanGenerality_norm2}
\title{Network normalised mean Generality -- following Williams and Martinez 2000}
\usage{
MeanGenerality_norm2(M)
}
\arguments{
\item{M}{a square \code{matrix} representing an adjacency newtork}
}
\description{
Network normalised mean Generality -- following Williams and Martinez 2000
}
