% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nu17May_Ceres.R
\name{MeanVulnerability}
\alias{MeanVulnerability}
\title{Network mean Vulnerability}
\usage{
MeanVulnerability(M)
}
\arguments{
\item{M}{a square \code{matrix} representing an adjacency newtork}
}
\description{
Network mean Vulnerability
}
