% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nu17May_Ceres.R
\name{MeanVulnerability_norm}
\alias{MeanVulnerability_norm}
\title{Network normalised mean Vulnerability}
\usage{
MeanVulnerability_norm(M)
}
\arguments{
\item{M}{a square \code{matrix} representing an adjacency newtork}
}
\description{
Network normalised mean Vulnerability
}
