% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nu17May_Ceres.R
\name{MeanVulnerability_norm2}
\alias{MeanVulnerability_norm2}
\title{Network normalised mean Vulnerability -- following Williams and Martinez 2000}
\usage{
MeanVulnerability_norm2(M)
}
\arguments{
\item{M}{a square \code{matrix} representing an adjacency newtork}
}
\description{
Network normalised mean Vulnerability -- following Williams and Martinez 2000
}
