% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkProperties.R
\name{netw.metrics}
\alias{netw.metrics}
\title{CALCULATE NETWORK METRICS
Function to calculate several network metrics}
\usage{
netw.metrics(web = NULL, normalise = FALSE, verbose = TRUE)
}
\arguments{
\item{web}{a square \code{matrix} representing an adjacency matrix.}

\item{normalise}{use normalised version of generality, vulnerability (and their
standard deviations)?}

\item{verbose}{print messages?}
}
\description{
CALCULATE NETWORK METRICS
Function to calculate several network metrics
}
