% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksBetadiversity.R
\name{networkTempBetaDiv.master}
\alias{networkTempBetaDiv.master}
\title{TAXONOMIC BETA-DIVERSITY FUNCTION FOR PIXEL BY SPP MATRICES}
\usage{
networkTempBetaDiv.master(pixXsppMat1, pixXsppMat2, method = "all", ...)
}
\arguments{
\item{pixXsppMat1}{is a data.table of pixel id's (column "PAGENAME") and species (other columns)
filled with 0s (absences from pixel network) and 1s (presences in pixel network)}

\item{pixXsppMat2}{same as \code{pixXsppMat1}. The network compositions to be compared with
\code{pixXsppMat1}}

\item{method}{is the method used to calculate beta diversity. Options are Sorensen's ("sorensen")
and Simpson's ("simpson") beta-diversity and their nested component ("nested") (see Baselga 2010),
the Jaccard dissimilarity index ("jaccard"), a multiplicative decomposition of alpha/gamma
diversity (where alpha and gamma are calculated as inverse-Simpson; "decompSimpson"),
all options ("all"), or any combination of them. Options "sorensen", "simpson", "nested", "jaccard",
"decompSimpson" calculate species beta-diversity (i.e., turnover). Defaults to "all"}
}
\value{
a data.table of beta-diversity indices per pixel ID (column "PAGENAME")
}
\description{
calculates beta diversity as the spp turnover between pairs of networks
   defined in two pixel by species matrices (containing species presences absences from
   networks in each pixel). Comparisons are made per pixel, between the two matrices,
   belonging from two different points in time, or two scenarios.
}
