% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksBetadiversity.R
\name{networkTempBetaDiv}
\alias{networkTempBetaDiv}
\title{TAXONOMIC BETA-DIVERSITY FUNCTION
calculates beta diversity as the spp turnover between two networks.}
\usage{
networkTempBetaDiv(web1, web2, dietcat, method = "all", mode = "composition")
}
\arguments{
\item{web1}{first adjency matrix (species network)}

\item{web2}{second adjency matrix (species network)}

\item{dietcat}{diet categories (i.e., lowest trophic level nodes that are ubiquitous
across all networks)}

\item{method}{is the method used to calculate beta diversity in 'composition' \code{mode}.
Options are Sorensen's ("sorensen") and Simpson's ("simpson") beta-diversity
and their nested component ("nested") (see Baselga 2010), the Jaccard
dissimilarity index ("jaccard"), a multiplicative decomposition of alpha/gamma
diversity (where alpha and gamma are calculated as inverse-Simpson; "decompSimpson"),
all options ("all"), or any combination of them. All these options
calculate beta-diversity (i.e., turnover) in terms of species composition.
Defaults to "all"}

\item{mode}{can be "decompHills", "pairwiseHills" or "composition". "decompHills" and "pairwiseHills"
use the \code{econetworks} package to calculate beta-diversity based on
species and link composition, either through alpha/gamma decomposition
("decompHills") or by calculating pairwise beta-diversity ("pairwiseHills").
In both cases, diversity is calculated using a Hills numbers approach. "composition" calculates
beta-diversity based on species composition only, using one of \code{method}.}
}
\description{
TAXONOMIC BETA-DIVERSITY FUNCTION
calculates beta diversity as the spp turnover between two networks.
}
