% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nu17May_Ceres.R
\name{NormalisedVulnerability}
\alias{NormalisedVulnerability}
\title{Node normalised Vulnerability}
\usage{
NormalisedVulnerability(M)
}
\arguments{
\item{M}{a square \code{matrix} representing an adjacency newtork}
}
\description{
Node normalised Vulnerability
}
