% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nu17May_Ceres.R
\name{NumberOfBasal}
\alias{NumberOfBasal}
\title{Number of basal nodes}
\usage{
NumberOfBasal(M)
}
\arguments{
\item{M}{a square \code{matrix} representing an adjacency newtork}
}
\description{
Number of basal nodes
}
