% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nu17May_Ceres.R
\name{NumberOfIntermediate}
\alias{NumberOfIntermediate}
\title{Number of intermediate consumers}
\usage{
NumberOfIntermediate(M)
}
\arguments{
\item{M}{a square \code{matrix} representing an adjacency newtork}
}
\description{
Number of intermediate consumers
}
