% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nu17May_Ceres.R
\name{NumberOfTop}
\alias{NumberOfTop}
\title{Number of top consumers}
\usage{
NumberOfTop(M)
}
\arguments{
\item{M}{a square \code{matrix} representing an adjacency newtork}
}
\description{
Number of top consumers
}
