% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nu17May_Ceres.R
\name{OutDegree}
\alias{OutDegree}
\title{Out-degree per node}
\usage{
OutDegree(M)
}
\arguments{
\item{M}{a square \code{matrix} representing an adjacency newtork}
}
\description{
Out-degree per node
}
