% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialTools.R
\name{outerBuffer}
\alias{outerBuffer}
\title{DRAW CONVEX HULL AROUND POLYGON}
\usage{
outerBuffer(x)
}
\arguments{
\item{x}{a \code{SpatialPolygons}, or a \code{SpatialPolygonsDF} object
NOTE: this function as been passed to amc.}
}
\value{
a convex hull polygon
}
\description{
Draws a convex hull around vertice points of a polygon \code{shapefile}.
}
