% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksDataPrep.R
\name{pixXspp_function}
\alias{pixXspp_function}
\title{CONVERT LIST OF WEBS/SPP IN PIXELS INTO A PIXELS x SSP MATRIX}
\usage{
pixXspp_function(list.obj = NULL, is.web = TRUE)
}
\arguments{
\item{list.obj}{is a list of webs (if is.web = TRUE), or of vectors of species,
per pixel that is to be converted into a matrix of pix X spp}

\item{is.web}{is \code{list.obj} a list of webs?}
}
\description{
CONVERT LIST OF WEBS/SPP IN PIXELS INTO A PIXELS x SSP MATRIX
}
