% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hypervolumes_functions.R
\name{plotHypervolumes3D}
\alias{plotHypervolumes3D}
\title{PLOT 3D HYPERVOLUMES WITH ORDINATION LOADINGS AND POST-HOC FITTED VECTORS}
\usage{
plotHypervolumes3D(
  HVlist,
  loadings_coords = NULL,
  PHvect_coords = NULL,
  loadings_labels = NULL,
  PHvect_labels = NULL,
  loadings_col = "grey20",
  cex.label = 1,
  lwd = 1,
  PHvect_col = "darkgreen",
  ...
)
}
\arguments{
\item{HVlist}{a \code{HypervolumeList}}

\item{loadings_coords}{a data.table or data.frame of ordination factor
loadings for the 3 axes to be plotted}

\item{PHvect_coords}{a data.table or data.frame of post-hoc fitted vectors
loadings for the 3 axes to be plotted}

\item{loadings_labels}{character vector of labels for each ordination factor.
Defaults to \code{row.names(loadings_coords)}}

\item{PHvect_labels}{character vector of labels for each post-hoc fitted vector.
Defaults to \code{row.names(PHvect_coords)}}

\item{loadings_col}{a colour for ordination factor arrows and labels.}

\item{cex.label}{passed to \code{basicPlotteR::addTextLabels}}

\item{lwd}{passed to \code{scatterplot3d::scatterplot3d}}

\item{PHvect_col}{a colour for post-hoc fitted vector arrows and labels.}

\item{...}{arguments passed to \code{.plot.HypervolumeListEdited}}
}
\description{
Plots two 3D hypervolumes that are based on ordination scores, showing
variable loadings and vectors fitted post-hoc to the ordination
}
