% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkProperties.R
\name{potential.inter}
\alias{potential.inter}
\title{DETECT INTERACTIONS FUNCTION}
\usage{
potential.inter(metaweb = NULL, SPPCODE = NULL, MODE = NULL, HELP = TRUE)
}
\arguments{
\item{metaweb}{a square \code{matrix} representing an adjacency metaweb}

\item{SPPCODE}{one or more species IDs}

\item{MODE}{one of "SUBSET", "EATS" or "EATEN". Use \code{HELP = TRUE} for a
description of each option}

\item{HELP}{print description of \code{MODE}?}
}
\description{
Function to detect potential prey/predators
}
