% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepCorrTable.R
\name{prepCorrTable}
\alias{prepCorrTable}
\title{LATEX/RMARKDOWN CORRELATION TABLES}
\usage{
prepCorrTable(x = NULL, corTable = NULL, pTable = NULL, method = "pearson")
}
\arguments{
\item{x}{a matrix or any object compatible with as.matrix, from where the covariance matrix will be calculated.
alternatively, corTable and pTable matrices can be supplied.}

\item{corTable}{matrix of correlation values obtained from e.g. \code{Hmisc::rcorr(...)$r}.
If supplied, x will be ignored}

\item{pTable}{matrix of pTable-values values obtained from e.g. \code{Hmisc::rcorr(...)$pTable}}

\item{method}{see \code{\link[Hmisc:rcorr]{Hmisc::rcorr()}} \code{type} argument.}
}
\description{
function to make a nice correlation table for latex/Rmd
usage example:
pTable <- rcorr()
tabl <- prepCorrTable(DT) \%>\% xtable::xtable(.)
print(tabl, type = "html.
Adapted from http://myowelt.blogspot.com/2008/04/beautiful-correlation-tables-in-r.html
}
