% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialTools.R
\name{prepKMZ2shapefile}
\alias{prepKMZ2shapefile}
\title{DOWNLOAD KMZ AND CONVERT TO SHAPEFILE}
\usage{
prepKMZ2shapefile(url, archive, destinationPath, overwrite = TRUE)
}
\arguments{
\item{url}{is a Google Drive URl}

\item{archive}{is the .zip file name where the .kmz is contained}

\item{destinationPath}{path is the path to where archive will downloaded to, and where the final shapefile will be saved.}

\item{overwrite}{passed to \code{googledrive::drive_download} and \code{sp::shapefile}}
}
\value{
a \code{shapefile}
}
\description{
DOWNLOAD KMZ AND CONVERT TO SHAPEFILE
}
