% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksUtils.R
\name{ras2matrix}
\alias{ras2matrix}
\title{CONVERT RASTER TO MATRIX OF PRESENCE/ABSENCES}
\usage{
ras2matrix(ras)
}
\arguments{
\item{ras}{a \code{raster} with species presence data.}
}
\description{
CONVERT RASTER TO MATRIX OF PRESENCE/ABSENCES
}
