% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialTools.R
\name{rasterizeCover}
\alias{rasterizeCover}
\title{FAST RASTERIZE ALL POLYGONS COVEVED BY POLYGON}
\usage{
rasterizeCover(rasterToMatch, shp, field, noDataVal = 0)
}
\arguments{
\item{rasterToMatch}{is the raster to use to burn polygon values to}

\item{shp}{is sf of \code{SpatialPolygonsDataFrame} object}

\item{field}{character. The field of shp to extract values from.
If NULL, 1s will be assigned to all cells that touch polygons}

\item{noDataVal}{Numeric, compatible with "Float64". value assigned to no data. Defaults to 0.}
}
\description{
A faster version of \code{rasterize} with the argument \code{getCover} that
for rasterizing polygon values that are touch a raster cell (rather than only it's center)
}
