% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialTools.R
\name{renameCleanSfFields}
\alias{renameCleanSfFields}
\title{RENAME AND SUBSET FIELDS IN SF OBJECT ACCORDING TO TABLE}
\usage{
renameCleanSfFields(sfObj, namesTable, rmNAs = TRUE)
}
\arguments{
\item{sfObj}{is an sf object from the sf package}

\item{namesTable}{is a two-column \code{data.frame} with the names to be replaced (1st column) and the new names (2nd column)
if there are NAs in the second columns, the original columns will be removed if \code{rmNAs == TRUE}}

\item{rmNAs}{determines whether columns with missing new names are removed or not. Defaults to \code{TRUE}}
}
\value{
an \code{sf} object
}
\description{
RENAME AND SUBSET FIELDS IN SF OBJECT ACCORDING TO TABLE
}
