% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksUtils.R
\name{replaceNAs.data.table}
\alias{replaceNAs.data.table}
\title{REPLACE NAs IN DATA.TABLE  --------------------------------
replaces NAs in a data.table with a value}
\usage{
replaceNAs.data.table(DT, value = 0L)
}
\arguments{
\item{DT}{is a data.table}

\item{value}{is the value to replace the NAs with. defaults to 0L}
}
\description{
REPLACE NAs IN DATA.TABLE  --------------------------------
replaces NAs in a data.table with a value
}
