% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMunging.R
\name{replaceNAs}
\alias{replaceNAs}
\title{REPLACE NAs BY ANOTHER VALUE IN A DATA.TABLE}
\usage{
replaceNAs(x, val = 0)
}
\arguments{
\item{x}{is a \code{data.table}}

\item{val}{is the value to replace NAs for. By default, it looks
for and replaces 0s}
}
\value{
a \code{data.table}
}
\description{
REPLACE NAs BY ANOTHER VALUE IN A DATA.TABLE
}
