% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksUtils.R
\name{robustnessMaps}
\alias{robustnessMaps}
\title{MAKE MAPS OF ROBUSTNESS METRICS}
\usage{
robustnessMaps(ras, mask = NULL, maskCol = NULL, factorLayer = FALSE)
}
\arguments{
\item{ras}{is a raster of a robustness metric (like mean robustness across scenarios)}

\item{mask}{is an optional mask to be plotted under the robustness metric (i.e. to show the full study area as dark grey surface)}

\item{maskCol}{is the colour used to fill mask values. Defaults to \code{"grey30"}}

\item{factorLayer}{is a boolean indicating whether values in \code{ras} should be treated \code{as.factor}}
}
\description{
MAKE MAPS OF ROBUSTNESS METRICS
}
