% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nu17May_Ceres.R
\name{SDGenerality}
\alias{SDGenerality}
\title{Network SD Generality}
\usage{
SDGenerality(M)
}
\arguments{
\item{M}{a square \code{matrix} representing an adjacency newtork}
}
\description{
Network SD Generality
}
