% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_nu17May_Ceres.R
\name{SDVulnerability_norm}
\alias{SDVulnerability_norm}
\title{Network normalised SD Vulnerability}
\usage{
SDVulnerability_norm(M)
}
\arguments{
\item{M}{a square \code{matrix} representing an adjacency newtork}
}
\description{
Network normalised SD Vulnerability
}
