% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialTools.R
\name{sfRmDupNACols}
\alias{sfRmDupNACols}
\title{CLEAN SF OBJECT FROM DUPLICATED COLUMNS}
\usage{
sfRmDupNACols(sfObj)
}
\arguments{
\item{sfObj}{is an sf object from the sf package}
}
\value{
an \code{sf} object without duplicated columns
}
\description{
this functions removes potentially duplicated data columns in an sf object
}
