% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkProperties.R
\name{spp.centrality}
\alias{spp.centrality}
\title{SPECIES CENTRALITY FUNCTION}
\usage{
spp.centrality(
  networkList,
  sppExclude,
  metric = c("degree", "betweenness", "eigenvector"),
  ...
)
}
\arguments{
\item{networkList}{a named \code{list} of networks, even if containing a single network}

\item{sppExclude}{list of species to exclude from degree centrality calculations
(all other metrics are calcualted on the full network)}

\item{metric}{centrality metric to calculate. Either "degree"
(using \code{igraph::degree(network, mode = "all", v = spp)}),
"betweenness" (using \code{igraph::betweenness(network, directed = TRUE)})
"eigenvector" centrality (using \code{eigen_centrality(graph = network, directed = FALSE, scale = FALSE)}).
(see Bauer et al 2010, Ecological Complexity).}

\item{...}{passed to reproducible::Cache}
}
\description{
Function to calculate the centrality of a list of spp within a network, or a list of networks
network can be a single network or a list of networks to extract spp centrality values from
if network is a list, then species also needs to be a list of equal length
}
