% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksBuildLocalWebs.R
\name{SppRm_localweb}
\alias{SppRm_localweb}
\title{Function to build local networks according to spp removal scenarios}
\usage{
SppRm_localweb(
  metaweb = NULL,
  SPPCODE = NULL,
  SPP.HAB = NULL,
  PIX.HAB = NULL,
  SPP.EXT = NULL,
  EXT.TRSH = NULL,
  dietcat = NULL,
  HELP = TRUE
)
}
\arguments{
\item{metaweb}{a square \code{matrix} representing an adjacency metaweb}

\item{SPPCODE}{one or more species IDs}

\item{SPP.HAB}{a spp x habitat \code{matrix}}

\item{PIX.HAB}{a pix x habitat \code{matrix}}

\item{SPP.EXT}{vector of spp to be primarily extinct}

\item{EXT.TRSH}{threshold of lost prey that determines spp extinction. Varies from 0 to 1,
increasing with predator sensitivity (e.g. 0.9 = spp dies if it is left with 90 \\\% of original prey)}

\item{dietcat}{diet categories (i.e., lowest trophic level nodes that are ubiquitous across all networks)}

\item{HELP}{print help}
}
\description{
Function to build local networks according to spp removal scenarios
}
