% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySE.R
\name{summarySE}
\alias{summarySE}
\title{Summarize data}
\usage{
summarySE(
  data = NULL,
  measurevar,
  groupvars = NULL,
  na.rm = FALSE,
  conf.interval = 0.95,
  .drop = TRUE
)
}
\arguments{
\item{data}{a data frame.}

\item{measurevar}{the name of a column that contains the variable to be summarized}

\item{groupvars}{a vector containing names of columns that contain grouping variables}

\item{na.rm}{a boolean that indicates whether to ignore NA's, defaults to FALSE.}

\item{conf.interval}{the percent range of the confidence interval (default is 95\%)}

\item{.drop}{passed to \code{ddply}}
}
\value{
A \code{data.frame} of summary statistics of each variable
 by grouping variables
}
\description{
Gives count, mean, standard deviation, standard error of the mean, and confidence interval (default 95\%).
}
