% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkProperties.R
\name{tlstats}
\alias{tlstats}
\title{CALCULATE TROPHIC LEVEL STATISTICS
Function to calculate trophic level stats}
\usage{
tlstats(web = NULL, dietcat = dietcat)
}
\arguments{
\item{web}{a square \code{matrix} representing an adjacency matrix.}

\item{dietcat}{diet categories (i.e., lowest trophic level nodes that are ubiquitous
across all networks)}
}
\description{
CALCULATE TROPHIC LEVEL STATISTICS
Function to calculate trophic level stats
}
