% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidation.R
\name{updateModelCached}
\alias{updateModelCached}
\title{CACHE-COMPATIBLE MODEL UPDATE}
\usage{
updateModelCached(..., cacheObj1 = NULL, cacheObj2 = NULL)
}
\arguments{
\item{...}{arguments passed to \code{update}}

\item{cacheObj1}{an object used by Cache for digesting, to avoid digesting
the (potentially) large data arguments e.g.: model coefficients and a
sample of a column drawn with a set seed.}

\item{cacheObj2}{an object used by Cache for digesting, to avoid digesting
the (potentially) large data arguments e.g.: model coefficients and a
sample of a column drawn with a set seed.}
}
\description{
to allow caching without digesting the large data table and model
}
