% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialTools.R
\name{validateGeomsSf}
\alias{validateGeomsSf}
\title{VALIDATE GEOMETRIES}
\usage{
validateGeomsSf(sfObj, dim)
}
\arguments{
\item{sfObj}{is an sf object from the sf package}

\item{dim}{is used for faster caching}
}
\value{
an \code{sf} object with validated geometries. Or
or an error if geometries could not be validated.
}
\description{
Checks for corrupt geometries in an \code{sf} object
and validates them using \code{sf::st_make_valid}
}
