% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialTools.R
\name{vector2binmatrix}
\alias{vector2binmatrix}
\title{VECTOR TO BINARY MATRIX}
\usage{
vector2binmatrix(x)
}
\arguments{
\item{x}{is a vector. Values must be coercible to character}
}
\value{
a presence/absence \code{matrix}
}
\description{
converts a vector of values into a binary "presence/absence" matrix
}
