% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksDataPrep.R
\name{whois}
\alias{whois}
\title{----------------------------------------------
NETWORKS DATA PROCESSING FUNCTIONS
----------------------------------------------
IDENTIFY SPP BY CODE}
\usage{
whois(SPPCODE = NULL, SPPNAME = NULL)
}
\arguments{
\item{SPPCODE}{one or more species IDs}

\item{SPPNAME}{Latin species name}
}
\description{
Function to Identify a spp by the code.
Author: João Braga 2016
}
