% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkProperties.R
\name{calcNetMeanDistances}
\alias{calcNetMeanDistances}
\title{CALCULATE MEAN BETA DIVERSITY (PAIRWISE DISTANCE)
Summarizes network pairwise distances into mean distance from a
focal network to others}
\usage{
calcNetMeanDistances(distObj)
}
\arguments{
\item{distObj}{a \code{dist} object calculated using \code{\link[econetwork:disPairwise]{econetwork::disPairwise()}}}
}
\description{
CALCULATE MEAN BETA DIVERSITY (PAIRWISE DISTANCE)
Summarizes network pairwise distances into mean distance from a
focal network to others
}
