% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksDataPrep.R
\name{compileResultsMasterScen}
\alias{compileResultsMasterScen}
\title{LOAD SCENARIO MASTER MATRICES (SPP X PIXEL MATRICES) FUNCTION
compiles simulation results and saves the data.table}
\usage{
compileResultsMasterScen(
  file.ls,
  res.dir,
  out.dir,
  quant,
  dietcat,
  useCache = NULL,
  cacheRepo = NULL
)
}
\arguments{
\item{file.ls}{is the list of file.paths for the master matrices
#' @template res.dir}

\item{out.dir}{is the directory where outputs will be saved}

\item{quant}{is the quantile threshold of extinction chosen}

\item{dietcat}{diet categories (i.e., lowest trophic level nodes that are ubiquitous
across all networks)}

\item{useCache}{is NULL, but will default to TRUE if the argument is not defined in the parent.frame()}

\item{cacheRepo}{is NULL, but will default to temp.dir() if the argument is not defined in the parent.frame()}
}
\description{
LOAD SCENARIO MASTER MATRICES (SPP X PIXEL MATRICES) FUNCTION
compiles simulation results and saves the data.table
}
