% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidation.R
\name{crossValidFunction}
\alias{crossValidFunction}
\title{----------------------------------------
K-FOLD CROSS VALIDATION FUNCTIONS
Ceres June 03 2020
----------------------------------------
CROSS-VALIDATION FUNCTION}
\usage{
crossValidFunction(
  fullDT,
  statsModel,
  origData,
  k = 4,
  idCol,
  parallel = FALSE,
  cacheObj1 = NULL,
  cacheObj2 = NULL,
  cacheArgs = NULL,
  level = NULL,
  ...
)
}
\arguments{
\item{fullDT}{\code{data.table} with full dataset}

\item{statsModel}{the statistical model to validate. Only works with gamlss models}

\item{k}{integer with number of chunks that the data should be partioned in}

\item{idCol}{column with pixel/observation IDs (optional)}

\item{parallel}{logical. Uses \code{future.apply::future_lapply} to parallelise
model fitting across the k-folds, using \code{plan(multiprocess)}. Defauls to FALSE}

\item{cacheArgs}{a named \code{list} of arguments passed to inner \code{Cache} calls}

\item{...}{further arguments passed to \code{future::plan}}

\item{origDataVars}{the data used to fit the statsModule, needs to be passed to \code{gamlss::predictAll}
 (it may not be able to access it) but also to make sure newdata in \code{gamlss::predictAll}
has the same variables (even if they're not used in the model)}

\item{cacheObj1/2}{an object used by \code{reproducible::Cache} for digesting,
to avoid digesting the (potentially) large data arguments}
}
\description{
----------------------------------------
K-FOLD CROSS VALIDATION FUNCTIONS
Ceres June 03 2020
----------------------------------------
CROSS-VALIDATION FUNCTION
}
