% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidation.R
\name{.calcMeanBEINF}
\alias{.calcMeanBEINF}
\title{CALCULATION OF THE MEAN FROM A BETA-INFLATED DISTRIBUTION
adapted from `gamlss::meanBEINF`}
\usage{
.calcMeanBEINF(mu, nu, tau)
}
\arguments{
\item{mu}{vector of values for the mu parameter of
the Beta-inflated distribution}

\item{nu}{vector of values for the nu parameter of
the Beta-inflated distribution}

\item{tau}{vector of values for the tau parameter of
the Beta-inflated distribution}
}
\description{
CALCULATION OF THE MEAN FROM A BETA-INFLATED DISTRIBUTION
adapted from `gamlss::meanBEINF`
}
