% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hypervolumes_functions.R
\name{HVordination}
\alias{HVordination}
\title{CALCULATE ORDINATION FOR HYPERVOLUME COMPUTATION}
\usage{
HVordination(
  datatable,
  HVidvar,
  init.vars = NULL,
  ordination = "PCA",
  noAxes = NULL,
  plotOrdi = TRUE,
  outputs.dir = NULL,
  file.suffix = NULL,
  saveOrdi = TRUE
)
}
\arguments{
\item{datatable}{is a \code{data.table} contained the raw data for the two hypervolumes}

\item{HVidvar}{the name or number of the column (in \code{HVdata1/2}) containing the ID of each hypervolume.}

\item{init.vars}{is a vector with the variable indices (column numbers) to use. Defaults to NULL
and all variables are used - ATTENTION the user should not have more than 8 variables unless
using an ordination to reduce dimensionality. See Blonder et al 2014.}

\item{ordination}{determines whether hypervolumes will be calculated on the raw data ("none") of on
ordination factor scores (one of "PCA", "HillSmith", "dudi.mix"). Defaults to using a PCA.}

\item{noAxes}{determines the number of axes/columns to use for hypervolume calculation}

\item{plotOrdi}{activates/desactivates plotting for ordinations - plots are saved to PDFs not plotted interactively}

\item{outputs.dir}{is the directory to store results}

\item{file.suffix}{is a character string used as a suffix in file names}

\item{saveOrdi}{activates/desactivates saving ordinations outputs}
}
\value{
a \code{list} of points used to build hypervolumes and final number of axes.
}
\description{
Calculates an ordination across two datasets prior to calculating two hypervolumes (one from each dataset)
}
