% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksDataPrep.R
\name{loadResultsSpaceBetaDiv}
\alias{loadResultsSpaceBetaDiv}
\title{LOAD SPATIAL BETA DIVERSITY RESULTS FUNCTION}
\usage{
loadResultsSpaceBetaDiv(bl.dir, res.dir, out.dir, quant)
}
\arguments{
\item{bl.dir}{is the directory where baseline simulation results were saved}

\item{res.dir}{is the directory where scenario simulation results were saved}

\item{out.dir}{is the directory where outputs will be saved}

\item{quant}{is the quantile threshold of extinction chosen}
}
\description{
compiles calculations of spatial beta diversity results and saves the data.table
}
