% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksDataPrep.R
\name{loadResultsTempBetaDiv}
\alias{loadResultsTempBetaDiv}
\title{LOAD TEMPORAL BETA DIVERSITY RESULTS FUNCTION}
\usage{
loadResultsTempBetaDiv(res.dir, out.dir, quant)
}
\arguments{
\item{res.dir}{is the directory where scenario simulation results were saved}

\item{out.dir}{is the directory where the output data.table will be saved as an RData file}

\item{quant}{is the quantile threshold of extinction chosen}
}
\description{
compiles calculations of temporal beta diversity results and saves the data.table
}
