% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networksDataPrep.R
\name{loadResultsTempBetaDiv}
\alias{loadResultsTempBetaDiv}
\title{LOAD TEMPORAL BETA DIVERSITY RESULTS FUNCTION}
\usage{
loadResultsTempBetaDiv(res.dir, out.dir, quant)
}
\arguments{
\item{out.dir}{is the directory where outputs will be saved}

\item{quant}{is the quantile threshold of extinction chosen}
}
\description{
compiles calculations of temporal beta diversity results and saves the data.table
}
\details{
#' @template res.dir
}
